+++ CHANGES.rst
@@ -5,7 +5,7 @@ kotori-mqtt CHANGES
 develop
 -------
 - upgrade foundation libraries: Twisted, Autobahn, Crossbar
-
+- add frontend foundation based on Pyramid web framework
 
 2015-05-21 0.2.2
 ----------------